+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project are documented here. The format of this docu
 
 ## [Unreleased]
 
-### Chaned
+### Added
+
+- Password hashers in `starlette_auth_toolkit.passwords`.
+
+### Changed
 
 - **BREAKING**: base backends are now under `.base.backends` instead of `.backends`.